[fix]coma/indep/mf
