+++ models/independent_ddpg.py
@@ -12,78 +12,85 @@ class IndependentDDPG(Model):
 
     def __init__(self, args, target_net=None):
         super(IndependentDDPG, self).__init__(args)
-        self.rl = DDPG(self.args)
         self.construct_model()
         self.apply(self.init_weights)
         if target_net != None:
             self.target_net = target_net
             self.reload_params_to_target()
         self.Transition = namedtuple('Transition', ('state', 'action', 'reward', 'next_state', 'done', 'last_step'))
+        self.rl = DDPG(self.args)
 
-    def unpack_data(self, batch):
-        batch_size = len(batch.state)
-        rewards = cuda_wrapper(torch.tensor(batch.reward, dtype=torch.float), self.cuda_)
-        last_step = cuda_wrapper(torch.tensor(batch.last_step, dtype=torch.float).contiguous().view(-1, 1), self.cuda_)
-        done = cuda_wrapper(torch.tensor(batch.done, dtype=torch.float).contiguous().view(-1, 1), self.cuda_)
-        actions = cuda_wrapper(torch.tensor(np.stack(list(zip(*batch.action))[0], axis=0), dtype=torch.float), self.cuda_)
-        state = cuda_wrapper(prep_obs(list(zip(batch.state))), self.cuda_)
-        next_state = cuda_wrapper(prep_obs(list(zip(batch.next_state))), self.cuda_)
-        return (rewards, last_step, done, actions, state, next_state)
 
     def construct_policy_net(self):
+        # TODO: fix policy params update
+        action_dicts = []
         if self.args.shared_parameters:
             l1 = nn.Linear(self.obs_dim, self.hid_dim)
             l2 = nn.Linear(self.hid_dim, self.hid_dim)
             a = nn.Linear(self.hid_dim, self.act_dim)
-            self.action_dict = nn.ModuleDict( {'layer_1': nn.ModuleList( [ l1 for _ in range(self.n_) ] ),\
-                                               'layer_2': nn.ModuleList( [ l2 for _ in range(self.n_) ] ),\
-                                               'action_head': nn.ModuleList( [ a for _ in range(self.n_) ] )
-                                              }
-                                            )
+            for i in range(self.n_):
+                action_dicts.append(nn.ModuleDict( {'layer_1': l1,\
+                                                    'layer_2': l2,\
+                                                    'action_head': a
+                                                    }
+                                                 )
+                                   )
         else:
-            self.action_dict = nn.ModuleDict( {'layer_1': nn.ModuleList( [ nn.Linear(self.obs_dim, self.hid_dim) for _ in range(self.n_) ] ),\
-                                               'layer_2': nn.ModuleList( [ nn.Linear(self.hid_dim, self.hid_dim) for _ in range(self.n_) ] ),\
-                                               'action_head': nn.ModuleList( [ nn.Linear(self.hid_dim, self.act_dim) for _ in range(self.n_) ] )
-                                              }
-                                            )
+            for i in range(self.n_):
+                action_dicts.append(nn.ModuleDict( {'layer_1': nn.Linear(self.obs_dim, self.hid_dim),\
+                                                    'layer_2': nn.Linear(self.hid_dim, self.hid_dim),\
+                                                    'action_head': nn.Linear(self.hid_dim, self.act_dim)
+                                                    }
+                                                  )
+                                   )
+        self.action_dicts = nn.ModuleList(action_dicts)
 
     def construct_value_net(self):
+        # TODO: policy params update
+        value_dicts = []
         if self.args.shared_parameters:
-            l1 = nn.Linear( (self.obs_dim+self.act_dim), self.hid_dim )
+            l1 = nn.Linear(self.obs_dim+self.act_dim, self.hid_dim )
             l2 = nn.Linear(self.hid_dim, self.hid_dim)
             v = nn.Linear(self.hid_dim, 1)
-            self.value_dict = nn.ModuleDict( {'layer_1': nn.ModuleList( [ l1 for _ in range(self.n_) ] ),\
-                                              'layer_2': nn.ModuleList( [ l2 for _ in range(self.n_) ] ),\
-                                              'value_head': nn.ModuleList( [ v for _ in range(self.n_) ] )
-                                             }
-                                           )
+            for i in range(self.n_):
+                value_dicts.append(nn.ModuleDict( {'layer_1': l1,\
+                                                   'layer_2': l2,\
+                                                   'value_head': v
+                                                  }
+                                                )
+                                  )
         else:
-            self.value_dict = nn.ModuleDict( {'layer_1': nn.ModuleList( [ nn.Linear( (self.obs_dim+self.act_dim), self.hid_dim ) for _ in range(self.n_) ] ),\
-                                              'layer_2': nn.ModuleList( [ nn.Linear(self.hid_dim, self.hid_dim) for _ in range(self.n_) ] ),\
-                                              'value_head': nn.ModuleList( [ nn.Linear(self.hid_dim, 1) for _ in range(self.n_) ] )
-                                             }
-                                           )
+            for i in range(self.n_):
+                value_dicts.append(nn.ModuleDict( {'layer_1': nn.Linear(self.obs_dim+self.act_dim, self.hid_dim ),\
+                                                   'layer_2': nn.Linear(self.hid_dim, self.hid_dim),\
+                                                   'value_head': nn.Linear(self.hid_dim, 1)
+                                                  }
+                                                )
+                                  )
+        self.value_dicts = nn.ModuleList(value_dicts)
 
     def construct_model(self):
         self.construct_value_net()
         self.construct_policy_net()
 
     def policy(self, obs, schedule=None, last_act=None, last_hid=None, info={}, stat={}):
+        # TODO: policy params update
         actions = []
         for i in range(self.n_):
-            h = torch.relu( self.action_dict['layer_1'][i](obs[:, i, :]) )
-            h = torch.relu( self.action_dict['layer_2'][i](h) )
-            a = self.action_dict['action_head'][i](h)
+            h = torch.relu( self.action_dicts[i]['layer_1'](obs[:, i, :]) )
+            h = torch.relu( self.action_dicts[i]['layer_2'](h) )
+            a = self.action_dicts[i]['action_head'](h)
             actions.append(a)
         actions = torch.stack(actions, dim=1)
         return actions
 
     def value(self, obs, act):
+        # TODO: policy params update
         values = []
         for i in range(self.n_):
-            h = torch.relu( self.value_dict['layer_1'][i]( torch.cat((obs[:,i,:],act[:,i,:]),dim=-1) ) )
-            h = torch.relu( self.value_dict['layer_2'][i](h) )
-            v = self.value_dict['value_head'][i](h)
+            h = torch.relu( self.value_dicts[i]['layer_1']( torch.cat((obs[:,i,:],act[:,i,:]),dim=-1)) 
+            h = torch.relu( self.value_dicts[i]['layer_2'](h) )
+            v = self.value_dicts[i]['value_head'](h)
             values.append(v)
         values = torch.stack(values, dim=1)
         return values
@@ -91,45 +98,3 @@ class IndependentDDPG(Model):
     def get_loss(self, batch):
         action_loss, value_loss, log_p_a = self.rl.get_loss(batch, self, self.target_net)
         return action_loss, value_loss, log_p_a
-
-    def train_process(self, stat, trainer):
-        info = {}
-        state = trainer.env.reset()
-        if self.args.reward_record_type is 'episode_mean_step':
-            trainer.mean_reward = 0
-            trainer.mean_success = 0
-        for t in range(self.args.max_steps):
-            state_ = cuda_wrapper(prep_obs(state).contiguous().view(1, self.n_, self.obs_dim), self.cuda_)
-            start_step = True if t == 0 else False
-            state_ = cuda_wrapper(prep_obs(state).contiguous().view(1, self.n_, self.obs_dim), self.cuda_)
-            action_out = self.policy(state_, info=info, stat=stat)
-            action = select_action(self.args, action_out, status='train', info=info)
-            _, actual = translate_action(self.args, action, trainer.env)
-            next_state, reward, done, debug = trainer.env.step(actual)
-            if isinstance(done, list): done = np.sum(done)
-            done_ = done or t==self.args.max_steps-1
-            trans = self.Transition(state,
-                                    action.cpu().numpy(),
-                                    np.array(reward),
-                                    next_state,
-                                    done,
-                                    done_
-                                   )
-            self.transition_update(trainer, trans, stat)
-            success = debug['success'] if 'success' in debug else 0.0
-            trainer.steps += 1
-            if self.args.reward_record_type is 'mean_step':
-                trainer.mean_reward = trainer.mean_reward + 1/trainer.steps*(np.mean(reward) - trainer.mean_reward)
-                trainer.mean_success = trainer.mean_success + 1/trainer.steps*(success - trainer.mean_success)
-            elif self.args.reward_record_type is 'episode_mean_step':
-                trainer.mean_reward = trainer.mean_reward + 1/(t+1)*(np.mean(reward) - trainer.mean_reward)
-                trainer.mean_success = trainer.mean_success + 1/(t+1)*(success - trainer.mean_success)
-            else:
-                raise RuntimeError('Please enter a correct reward record type, e.g. mean_step or episode_mean_step.')
-            stat['mean_reward'] = trainer.mean_reward
-            stat['mean_success'] = trainer.mean_success 
-            if done_:
-                break
-            state = next_state
-        stat['turn'] = t+1
-        trainer.episodes += 1
